+++ src/openapi.yaml
@@ -7779,6 +7779,7 @@ components:
           description: For jq, pass a string containing the filter expression
       required:
       - id
+      - name
       - service
     SourceEnum:
       enum:
