+++ src/openforms/variables/admin.py
@@ -9,8 +9,8 @@ class ServiceFetchConfigurationAdmin(admin.ModelAdmin):
         "service",
     ]
     list_display = [
-        "service",
         "name",
+        "service",
         "method",
         "path",
     ]
@@ -18,6 +18,7 @@ class ServiceFetchConfigurationAdmin(admin.ModelAdmin):
         "service",
     ]
     search_fields = [
+        "name",
         "service__api_root",
         "path",
     ]
